Simple camera manager implemented.
Player camera is assigned to the camera manager on start.

The camera automatically shows the area in the direction of movement of the player to a given length to show where the player is moving towards.

The camera movement only after the player moves within a given bounds from the centre of the viewport of the camera.